requirements_dev.txt
-r requirements.txt
model-bakery
pytest
pytest-django
#factory-man
black
django-debug-toolbar
django_extensions
# pylint
pylint-django
pylint-pytest
django-stubs[compatible-mypy]
# flake8
